vision constants
