+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/commands/custom/ExtendCommand.java
@@ -1,57 +1,15 @@
 package org.firstinspires.ftc.teamcode.commands.custom;
 
-import android.util.Log;
-import com.arcrobotics.ftclib.command.CommandBase;
-import javax.annotation.Nullable;
 import org.firstinspires.ftc.teamcode.subsystems.ExtensionSubsystem;
 
-public class ExtendCommand extends CommandBase {
-    ExtensionSubsystem extend;
-    double target;
-    @Nullable private Double powerMul = null;
-    private double oldExtensionPowerMul;
+public class ExtendCommand extends TimeoutCommand {
 
     /**
      * @param subsystem
-     * @param target in inches
+     * @param target    in inches
      */
     public ExtendCommand(ExtensionSubsystem subsystem, double target) {
-        this.extend = subsystem;
-        this.target = target;
-        addRequirements(subsystem);
+        super(new ExtendCommandInternal(subsystem, target), (int) subsystem.getReasonableExtensionMillis(target));
     }
 
-    /**
-     * @param subsystem
-     * @param target in inches
-     * @param powerMul the power multipler to feed into the subsystem, will reset to the previous
-     *     power multiplier once command finishes
-     */
-    public ExtendCommand(ExtensionSubsystem subsystem, double target, double powerMul) {
-        this(subsystem, target);
-        this.powerMul = powerMul;
-    }
-
-    @Override
-    public void initialize() {
-        if (powerMul != null) {
-            oldExtensionPowerMul = extend.getPowerMul();
-            extend.setPowerMul(powerMul);
-        }
-        extend.setTargetInches(target);
-        extend.setManualControl(false);
-    }
-
-    @Override
-    public boolean isFinished() {
-        return extend.isClose(target);
-    }
-
-    @Override
-    public void end(boolean wasInterrupted) {
-        if (powerMul != null) {
-            extend.setPowerMul(oldExtensionPowerMul);
-        }
-        Log.i("%9", "Extension to " + target);
-    }
 }
